+++ CHANGELOG.md
@@ -21,6 +21,7 @@ x.x.x
   - Change the default deployment type to `Scratch_2`.
   - Change the Lustre server version to `2.12`.
 - Add `lambda:ListTags` and `lambda:UntagResource` to `ParallelClusterUserRole` used by ParallelCluster API stack for cluster update.
+- Add `parallelcluster:cluster-name` tag to all resources created by ParallelCluster.
 - Add new configuration parameter `Scheduling/SlurmSettings/EnableMemoryBasedScheduling` to configure memory-based
   scheduling in Slurm.